Add rubocop and fixed code complexity
